# frontend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Kitco News API",
        "description": "API documentation for managing articles and AI-powered operations.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "Article": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "6408b6f1a43e3b3f89cd17e1"
                    },
                    "title": {
                        "type": "string",
                        "example": "AI in Blockchain"
                    },
                    "titleFr": {
                        "type": "string",
                        "example": "AI dans le blockchain"
                    },
                    "content": {
                        "type": "string",
                        "example": "Exploring the intersection of AI and blockchain..."
                    },
                    "contentFr": {
                        "type": "string",
                        "example": "Exploration de l'intersection entre l'IA et le blockchain..."
                    },
                    "author": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "publishedDate": {
                        "type": "string",
                        "example": "2024-01-01"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Gold",
                            "Crypto"
                        ]
                    },
                    "category": {
                        "type": "string",
                        "example": "crypto"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "draft"
                    },
                    "relatedCompanies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Goldcorp",
                            "Barrick Gold"
                        ]
                    },
                    "marketData": {
                        "type": "string",
                        "example": "Gold price: $1,800/oz"
                    },
                    "summary": {
                        "type": "string",
                        "example": "AI in blockchain is a new way to do things"
                    }
                },
                "required": [
                    "title",
                    "content",
                    "author"
                ]
            },
            "ArticleAiRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "6408b6f1a43e3b3f89cd17e1"
                    },
                    "pressRelease": {
                        "type": "string",
                        "example": "This is a fake press release"
                    },
                    "language": {
                        "type": "string",
                        "example": "english"
                    }
                }
            }
        }
    },
    "paths": {
        "/articles": {
            "get": {
                "summary": "Get all articles",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Filter by title"
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Filter by author"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        },
                        "description": "Number of results per page"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        },
                        "description": "Page number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create an article",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "summary": "Get article by ID",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Article ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update an article",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Article ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete an article",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Article ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Article deleted"
                    }
                }
            }
        },
        "/articles/generate": {
            "post": {
                "summary": "Generate an article using AI",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleAiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Generated article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}/translate": {
            "post": {
                "summary": "Translate an article using AI",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Article ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleAiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translated article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}/summarize": {
            "post": {
                "summary": "Summarize an article using AI",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Article ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleAiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Summarized article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}/tag": {
            "post": {
                "summary": "Generate tags for an article using AI",
                "tags": [
                    "Articles"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Article ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleAiRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tagged article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}